bank-service:reformat test for bank profit
