+++ bank-service/src/test/java/rs/banka4/bank_service/integration/SecuritiesControllerTests.java
@@ -1,8 +1,10 @@
 package rs.banka4.bank_service.integration;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static rs.banka4.bank_service.utils.AssetGenerator.STOCK_EX1_UUID;
 
 import java.math.BigDecimal;
+import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -12,11 +14,13 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.assertj.MockMvcTester;
 import rs.banka4.bank_service.domain.user.client.db.Client;
 import rs.banka4.bank_service.integration.generator.PortfolioGenerator;
+import rs.banka4.bank_service.integration.generator.UserGenerator;
 import rs.banka4.bank_service.repositories.*;
 import rs.banka4.bank_service.utils.AssetGenerator;
 import rs.banka4.bank_service.utils.ExchangeGenerator;
 import rs.banka4.bank_service.utils.ListingGenerator;
 import rs.banka4.rafeisen.common.currency.CurrencyCode;
+import rs.banka4.rafeisen.common.security.Privilege;
 import rs.banka4.testlib.integration.DbEnabledTest;
 import rs.banka4.testlib.utils.JwtPlaceholders;
 
@@ -41,7 +45,10 @@ public class SecuritiesControllerTests {
     private PortfolioGenerator portfolioGenerator;
     @Autowired
     private OrderRepository orderRepository;
-
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private UserGenerator userGen;
 
     @Test
     public void testGetMyPortfolioWithStock() {
@@ -843,5 +850,93 @@ public class SecuritiesControllerTests {
             .isLenientlyEqualTo(expectedJson);
     }
 
+    @Test
+    void testBankProfitEndpoint() {
+        Client client = portfolioGenerator.createTestClientDiffIf();
+        var acts = portfolioGenerator.createTestActuary();
+        var clients =
+            acts.stream()
+                .map(act -> {
+                    var u = userRepository.findById(act.getUserId());
+                    u.get()
+                        .setPrivileges(List.of(Privilege.AGENT));
+                    return userRepository.save(u.get());
+                })
+                .toList();
+        final var ber1 = ExchangeGenerator.makeBer1();
+        exchangeRepo.save(ber1);
+        AssetGenerator.makeExampleAssets()
+            .forEach(assetRepository::saveAndFlush);
+        var stock = securityRepository.findById(STOCK_EX1_UUID);
+        ListingGenerator.makeExampleListings(
+            stock.orElseThrow(),
+            ber1,
+            listingRepo,
+            listingHistoryRepo
+        );
+        var orderB =
+            portfolioGenerator.createDummyBuyOrder(
+                client,
+                stock.get(),
+                100,
+                BigDecimal.valueOf(40),
+                CurrencyCode.EUR
+            );
+        var orderS =
+            portfolioGenerator.createDummySellOrder(
+                client,
+                stock.get(),
+                200,
+                BigDecimal.valueOf(50),
+                CurrencyCode.EUR
+            );
+        orderS.setUser(clients.get(0));
+        orderB.setUser(clients.get(0));
+        orderRepository.saveAllAndFlush(List.of(orderB, orderS));
+        String jwtToken = "Bearer " + JwtPlaceholders.ADMIN_EMPLOYEE_TOKEN;
+        userGen.createEmployee(
+            e -> e.id(JwtPlaceholders.CLIENT_ID)
+                .email("bala@bla.com")
+                .username("bala")
+        );
+        String json = """
+            {
+              "content": [
+                {
+                  "profit": {
+                    "amount": 116747.94841735054008,
+                    "currency": "RSD"
+                  },
+                  "name": "John",
+                  "surname": "Doe",
+                  "position": "Developer"
+                },
+                {
+                  "profit": {
+                    "amount": 0,
+                    "currency": "RSD"
+                  },
+                  "name": "John",
+                  "surname": "Doe",
+                  "position": "Developer"
+                }
+              ],
+              "page": {
+                "size": 10,
+                "number": 0,
+                "totalElements": 2,
+                "totalPages": 1
+              }
+            }""";
+
+        mvc.get()
+            .uri("/stock/securities/bank/profit")
+            .header(HttpHeaders.AUTHORIZATION, jwtToken)
+            .contentType(MediaType.APPLICATION_JSON)
+            .assertThat()
+            .hasStatusOk()
+            .bodyJson()
+            .isLenientlyEqualTo(json);
+    }
 
 }
